@@ -1,6 +1,6 @@
 {
   "name": "@vue/server-renderer",
-  "version": "3.0.11",
+  "version": "3.1.0-beta.7",
   "description": "@vue/server-renderer",
   "main": "index.js",
   "types": "dist/server-renderer.d.ts",
@@ -28,10 +28,10 @@
   },
   "homepage": "https://github.com/vuejs/vue-next/tree/master/packages/server-renderer#readme",
   "peerDependencies": {
-    "vue": "3.0.11"
+    "vue": "3.1.0-beta.7"
   },
   "dependencies": {
-    "@vue/shared": "3.0.11",
-    "@vue/compiler-ssr": "3.0.11"
+    "@vue/shared": "3.1.0-beta.7",
+    "@vue/compiler-ssr": "3.1.0-beta.7"
   }
 }
