@@ -281,7 +281,7 @@ function convertLegacySlots(vnode: VNode): VNode {
       for (const key in slots) {
         const slotChildren = slots[key]
         slots[key] = () => slotChildren
-        slots[key]._nonScoped = true
+        slots[key]._ns = true /* non-scoped slot */
       }
     }
   }
